refactor(analytics): Changed mixpanel imports
