+++ src/shared/hooks/useLogout.ts
@@ -7,7 +7,7 @@ import { useAppDispatch } from 'redux/store';
 import { alerts, auth, workspaces } from 'redux/modules';
 import { deleteAccessTokenApi, deleteRefreshTokenApi } from 'modules/Auth/api';
 
-import { Mixpanel } from 'shared/utils';
+import { Mixpanel } from 'shared/utils/mixpanel';
 
 export const useLogout = () => {
   const dispatch = useAppDispatch();
